+++ src/main/java/com/exam/literaryplanner/domain/Board.java
@@ -1,8 +1,16 @@
 package com.exam.literaryplanner.domain;
 
-import jakarta.persistence.*;
 import java.time.LocalDateTime;
+
 import org.hibernate.annotations.CreationTimestamp;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Lob;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "boardT")
@@ -11,7 +19,7 @@ public class Board {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "b_idx")
-    private Long bIdx;
+    private Integer bIdx;
 
     @Column(name = "b_type", nullable = false, length = 10)
     private String bType; // NOTICE / FAQ
@@ -32,8 +40,8 @@ public class Board {
 
     public Board() {}
 
-    public Long getBIdx() { return bIdx; }
-    public void setBIdx(Long bIdx) { this.bIdx = bIdx; }
+    public Integer getBIdx() { return bIdx; }
+    public void setBIdx(Integer bIdx) { this.bIdx = bIdx; }
 
     public String getBType() { return bType; }
     public void setBType(String bType) { this.bType = bType; }
