+++ src/main/java/com/exam/literaryplanner/domain/PasswordResetToken.java
@@ -1,7 +1,16 @@
 package com.exam.literaryplanner.domain;
 
-import jakarta.persistence.*;
 import java.time.LocalDateTime;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "password_reset_token")
@@ -9,7 +18,7 @@ public class PasswordResetToken {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     // memberT(m_idx) FK
     @ManyToOne(fetch = FetchType.LAZY)
@@ -37,7 +46,7 @@ public class PasswordResetToken {
     }
 
     // getters/setters
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
 
     public Member getMember() { return member; }
     public void setMember(Member member) { this.member = member; }
