+++ src/main/java/com/exam/literaryplanner/repository/QnaRepository.java
@@ -1,18 +1,19 @@
 package com.exam.literaryplanner.repository;
 
-import com.exam.literaryplanner.domain.Qna;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.util.List;
+import com.exam.literaryplanner.domain.Qna;
 
-public interface QnaRepository extends JpaRepository<Qna, Long> {
+public interface QnaRepository extends JpaRepository<Qna, Integer> {
 
 	@Query("select q from Qna q where q.member.id = :memberId order by q.qRegDate desc")
-	List<Qna> findQnaByMemberId(@Param("memberId") Long memberId);
+	List<Qna> findQnaByMemberId(@Param("memberId") Integer memberId);
 
 	@Query("select count(q) from Qna q where q.qStatus = :status")
-	long countPendingByStatus(@Param("status") Integer status);
+	int countPendingByStatus(@Param("status") Integer status);
 
 }
